Move siteConfig.json
